Refactor codebase
