+++ src/Actions/allMeals.js
@@ -3,20 +3,18 @@ import {
   fetchRecipesError,
 } from './index';
 
-function fetchAllMeals(type) {
-  return dispatch => {
-    fetch(`https://www.themealdb.com/api/json/v1/1/filter.php?c=${type}`)
-      .then(response => response.json())
-      .then(response => {
-        if (response.error) {
-          throw (response.error);
-        }
-        dispatch(fetchRecipesSuccess(response.meals));
-      })
-      .catch(error => {
-        dispatch(fetchRecipesError(error));
-      });
-  };
-}
+const fetchAllMeals = type => dispatch => {
+  fetch(`https://www.themealdb.com/api/json/v1/1/filter.php?c=${type}`)
+    .then(res => res.json())
+    .then(res => {
+      if (res.error) {
+        throw (res.error);
+      }
+      dispatch(fetchRecipesSuccess(res.meals));
+    })
+    .catch(error => {
+      dispatch(fetchRecipesError(error));
+    });
+};
 
 export default fetchAllMeals;
